requirements.txt
# Core dependencies
ollama>=0.1.7
chromadb>=0.4.15
langchain>=0.1.0
langchain-community>=0.0.10

# Document processing
PyPDF2>=3.0.1
python-docx>=0.8.11
pandas>=2.0.0
openpyxl>=3.1.0
textract>=1.6.5
# Enhanced document processing
pdfplumber>=0.9.0
tabula-py>=2.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0

# Vector database and embeddings
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0

# Web interface
flask>=2.3.0
flask-cors>=4.0.0
jinja2>=3.1.0

# Terminal interface
rich>=13.0.0
prompt-toolkit>=3.0.0
click>=8.1.0

# Performance and utilities
tqdm>=4.65.0
psutil>=5.9.0
multiprocessing-logging>=0.3.4

# Database
sqlite3  # Built-in with Python
sqlalchemy>=2.0.0

# Configuration and logging
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
